+++ internal/sink/sse_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"log-listener/internal/render"
+	"github.com/homeend/log-listener/internal/render"
 )
 
 func TestSSEHubStreamsEvents(t *testing.T) {